+++ keanu-project/src/main/java/io/improbable/keanu/tensor/TypedINDArrayFactory.java
@@ -1,25 +1,24 @@
 package io.improbable.keanu.tensor;
 
-import io.improbable.keanu.tensor.dbl.Nd4jDoubleTensor;
 import org.nd4j.linalg.api.buffer.DataBuffer;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.factory.Nd4j;
 
 public class TypedINDArrayFactory {
 
-    public static INDArray create(double[] data, int[] shape, DataBuffer.Type bufferType) {
+    public static INDArray create(double[] data, long[] shape, DataBuffer.Type bufferType) {
         Nd4j.setDataType(bufferType);
         DataBuffer buffer = Nd4j.getDataBufferFactory().createDouble(data);
         return Nd4j.create(buffer, shape);
     }
 
-    public static INDArray create(int[] data, int[] shape, DataBuffer.Type bufferType) {
+    public static INDArray create(int[] data, long[] shape, DataBuffer.Type bufferType) {
         Nd4j.setDataType(bufferType);
         DataBuffer buffer = Nd4j.getDataBufferFactory().createDouble(data);
         return Nd4j.create(buffer, shape);
     }
 
-    public static INDArray valueArrayOf(int[] shape, double value, DataBuffer.Type bufferType) {
+    public static INDArray valueArrayOf(long[] shape, double value, DataBuffer.Type bufferType) {
         Nd4j.setDataType(bufferType);
         return Nd4j.valueArrayOf(shape, value);
     }
@@ -29,7 +28,7 @@ public class TypedINDArrayFactory {
         return Nd4j.scalar(scalarValue);
     }
 
-    public static INDArray ones(int[] shape, DataBuffer.Type bufferType) {
+    public static INDArray ones(long[] shape, DataBuffer.Type bufferType) {
         Nd4j.setDataType(bufferType);
         return Nd4j.ones(shape);
     }
@@ -39,7 +38,7 @@ public class TypedINDArrayFactory {
         return Nd4j.eye(n);
     }
 
-    public static INDArray zeros(int[] shape, DataBuffer.Type bufferType) {
+    public static INDArray zeros(long[] shape, DataBuffer.Type bufferType) {
         Nd4j.setDataType(bufferType);
         return Nd4j.zeros(shape);
     }
