+++ keanu-project/src/main/java/io/improbable/keanu/vertices/dbl/probabilistic/MultivariateGaussianVertex.java
@@ -23,7 +23,7 @@ public class MultivariateGaussianVertex extends DoubleVertex implements Probabil
      * @param mu         the mu of the Multivariate Gaussian
      * @param covariance the covariance matrix of the Multivariate Gaussian
      */
-    public MultivariateGaussianVertex(int[] shape, DoubleVertex mu, DoubleVertex covariance) {
+    public MultivariateGaussianVertex(long[] shape, DoubleVertex mu, DoubleVertex covariance) {
 
         checkValidMultivariateShape(mu.getShape(), covariance.getShape());
 
@@ -53,7 +53,7 @@ public class MultivariateGaussianVertex extends DoubleVertex implements Probabil
      * @param covariance the scale of the identity matrix
      */
     public MultivariateGaussianVertex(DoubleVertex mu, double covariance) {
-        this(mu, ConstantVertex.of(DoubleTensor.eye(mu.getShape()[0])).times(covariance));
+        this(mu, ConstantVertex.of(DoubleTensor.eye((int)mu.getShape()[0])).times(covariance));
     }
 
     public MultivariateGaussianVertex(double mu, double covariance) {
@@ -78,7 +78,7 @@ public class MultivariateGaussianVertex extends DoubleVertex implements Probabil
         return MultivariateGaussian.withParameters(mu.getValue(), covariance.getValue()).sample(mu.getShape(), random);
     }
 
-    private static int[] checkValidMultivariateShape(int[] muShape, int[] covarianceShape) {
+    private static long[] checkValidMultivariateShape(long[] muShape, long[] covarianceShape) {
         if (covarianceShape.length != 2
             || muShape.length != 2
             || covarianceShape[0] != covarianceShape[1]
